Baixei o JUnit e criei classes para testes.
