+++ Aula_C14_Maven/jogos.json
@@ -1,20 +1,1 @@
-[
-  {
-    "nome": "Persona 4 Golden",
-    "dataLancamento": "14/06/2012",
-    "genero": "RPG",
-    "duracao": 100
-  },
-  {
-    "nome": "Disco Elysium The Final Cut",
-    "dataLancamento": "30/03/2021",
-    "genero": "RPG/Detetive",
-    "duracao": 15
-  },
-  {
-    "nome": "Minecraft",
-    "dataLancamento": "18/11/2011",
-    "genero": "Sandbox",
-    "duracao": 200
-  }
-]
+[]
