--- Aula_C14_Maven/src/main/java/br/inatel/jogos/Jogos.java
@@ -1,19 +1,0 @@
-package br.inatel.jogos;
-
-public class Jogos {
-    private String nome;
-    private String dataLancamento;
-    private String genero;
-    private int duracao;
-
-    public Jogos(String nome, String dataLancamento, String genero, int duracao) {
-        this.nome = nome;
-        this.dataLancamento = dataLancamento;
-        this.genero = genero;
-        this.duracao = duracao;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-}
