+++ Aula_C14_Maven/src/test/java/br/inatel/test/MainTeste.java
@@ -1,0 +1,10 @@
+package br.inatel.test;
+
+import br.inatel.Jogos;
+import org.junit.Test;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class MainTeste {
+}
